+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Reduce PIN resend throttle from 5 minutes to 1 minute.
 - Rename `token_auths` database table to `pin_logins`.
 - Redirect to PIN verify page after sending PIN email.
+- Simplify PIN email: remove button, show PIN and login link as plain text.
 
 ## [6.1.1] - 2026-03-08